refactored auth dir into multiple files
